Tidyup Schema & creating list schema
